@@ -1,9 +1,10 @@
+# frozen_string_literal: true
+
 module SalsaLabs
   ##
   # Service object to pull back a collection of objects from the Salsa Labs API.
   ##
   class SalsaObjectsFetcher
-
     def initialize(filter_parameters = {}, credentials = {})
       @filter_parameters = SalsaLabs::ApiObjectParameterList.new(filter_parameters)
       @client = SalsaLabs::ApiClient.new(credentials)
@@ -23,32 +24,30 @@ module SalsaLabs
 
     def get_objects
       client.fetch('/api/getObjects.sjs', api_parameters)
-      #Note, this will return at most 500 records
-      #TODO, implement pagination
+      # Note, this will return at most 500 records
+      # TODO, implement pagination
     end
 
     def get_tagged_objects(tag)
-      tag_parameters = api_parameters.update({'tag'=>tag})
+      tag_parameters = api_parameters.update({ 'tag' => tag })
       client.fetch('/api/getTaggedObjects.sjs', tag_parameters)
     end
 
     def api_parameters
-      if filter_parameters
-        params = {'condition'=>filter_parameters.attributes.flat_map {|k,v| "#{k}=#{v}" }}
-      else
-        params = {} 
-      end
+      params = if filter_parameters
+                 { 'condition' => filter_parameters.attributes.flat_map { |k, v| "#{k}=#{v}" } }
+               else
+                 {}
+               end
 
       params.merge(object: @object_class.object_name)
     end
 
-
     def item_objects(url)
       item_nodes(url).map do |node|
         obj = @object_class.new(ApiObjectNode.new(node).attributes)
-        if obj.attributes['result'] == 'error'
-          raise SalsaLabs::Error.new(obj.attributes.inspect)
-        end
+        raise SalsaLabs::Error, obj.attributes.inspect if obj.attributes['result'] == 'error'
+
         obj
       end
     end
@@ -62,15 +61,13 @@ module SalsaLabs
     # SalsaLabs::Object creation.
     ##
     class ApiObjectNode
-
       def initialize(xml_element)
         @node = xml_element
       end
 
       def attributes
-        children.inject({}) do |memo, attribute|
+        children.each_with_object({}) do |attribute, memo|
           memo[attribute.name.downcase] = attribute.text
-          memo
         end
       end
 
@@ -81,7 +78,6 @@ module SalsaLabs
       def children
         node.children
       end
-
     end
   end
 end
