@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module SalsaLabs
-  VERSION = "0.1.2"
+  VERSION = '0.1.2'
 end
