@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require 'salsa_labs'
 require 'dotenv'
 require 'faker'
@@ -6,6 +8,7 @@ Dotenv.load
 
 RSpec.configure do |config|
   config.before(:suite) do
-    SANDBOX_CREDENTIALS = {email: ENV['SALSA_LABS_API_EMAIL'], password: ENV['SALSA_LABS_API_PASSWORD'], url: ENV['https://sandbox.salsalabs.com']}
+    SANDBOX_CREDENTIALS = { email: ENV['SALSA_LABS_API_EMAIL'], password: ENV['SALSA_LABS_API_PASSWORD'],
+                            url: ENV['https://sandbox.salsalabs.com'] }.freeze
   end
 end
