@@ -1,7 +1,8 @@
-require "spec_helper"
+# frozen_string_literal: true
+
+require 'spec_helper'
 
 describe SalsaLabs::ActionsFetcher do
-
   let(:client) do
     double(
       'API client',
@@ -9,24 +10,23 @@ describe SalsaLabs::ActionsFetcher do
     )
   end
 
-  describe "#fetch" do
+  describe '#fetch' do
     before(:each) do
       allow(SalsaLabs::ApiClient).to receive(:new) { client }
     end
 
-    it "calls the getObjects API endpoint" do
+    it 'calls the getObjects API endpoint' do
       SalsaLabs::ActionsFetcher.new.fetch
 
-      expect(client).to have_received(:fetch).
-        with('/api/getObjects.sjs', {"condition"=>[], object: 'action'})
+      expect(client).to have_received(:fetch)
+        .with('/api/getObjects.sjs', { 'condition' => [], object: 'action' })
     end
 
-    it "returns an array of SalsaLabs::Action objects" do
+    it 'returns an array of SalsaLabs::Action objects' do
       results = SalsaLabs::ActionsFetcher.new.fetch
 
       expect(results).to be_a(Array)
       expect(results.first).to be_a(SalsaLabs::Action)
     end
   end
-
 end
