@@ -1,7 +1,8 @@
-require "spec_helper"
+# frozen_string_literal: true
+
+require 'spec_helper'
 
 describe SalsaLabs::Supporter do
-
   let(:attributes) do
     {
       'supporter_key' => '31337',
@@ -34,7 +35,7 @@ describe SalsaLabs::Supporter do
   let(:supporter) { SalsaLabs::Supporter.new(attributes) }
 
   describe 'initialization' do
-    let(:attributes) { {:first_name => 'George'} }
+    let(:attributes) { { first_name: 'George' } }
     it 'should accept symbol keys' do
       expect(supporter.first_name).to eq('George')
     end
@@ -45,141 +46,141 @@ describe SalsaLabs::Supporter do
     end
   end
 
-  describe "#attributes" do
-    it "returns the attributes hash passed in to initialize" do
+  describe '#attributes' do
+    it 'returns the attributes hash passed in to initialize' do
       expect(supporter.attributes).to eq(attributes)
     end
   end
 
-  describe "#supporter_key" do
-    it "returns the suppoter_key attribute as an integer" do
-      expect(supporter.supporter_key).to eq(31337)
+  describe '#supporter_key' do
+    it 'returns the suppoter_key attribute as an integer' do
+      expect(supporter.supporter_key).to eq(31_337)
     end
   end
 
-  describe "#organization_key" do
-    it "returns the organization_key attribute as an integer" do
+  describe '#organization_key' do
+    it 'returns the organization_key attribute as an integer' do
       expect(supporter.organization_key).to eq(1234)
     end
   end
 
-  describe "#chapter_key" do
-    it "returns the chapter_key attribute as an integer" do
-      expect(supporter.chapter_key).to eq(90210)
+  describe '#chapter_key' do
+    it 'returns the chapter_key attribute as an integer' do
+      expect(supporter.chapter_key).to eq(90_210)
     end
   end
 
-  describe "#title" do
-    it "returns the title as an attribute" do
+  describe '#title' do
+    it 'returns the title as an attribute' do
       expect(supporter.title).to eq('Mr.')
     end
   end
 
-  describe "#first_name" do
-    it "returns the first name as an attribute" do
+  describe '#first_name' do
+    it 'returns the first name as an attribute' do
       expect(supporter.first_name).to eq('John')
     end
   end
 
-  describe "#first_name=" do
-    it "sets the first name as an attribute" do
+  describe '#first_name=' do
+    it 'sets the first name as an attribute' do
       supporter.first_name = 'Nathan'
       expect(supporter.first_name).to eq('Nathan')
     end
   end
 
-  describe "#mi" do
-    it "returns the middle initial as an attribute" do
+  describe '#mi' do
+    it 'returns the middle initial as an attribute' do
       expect(supporter.mi).to eq('Jacob')
     end
   end
 
-  describe "#last_name" do
-    it "returns the last_name as an attribute" do
+  describe '#last_name' do
+    it 'returns the last_name as an attribute' do
       expect(supporter.last_name).to eq('Jingleheimer Schmidt')
     end
   end
 
-  describe "#suffix" do
-    it "returns the suffix as an attribute" do
+  describe '#suffix' do
+    it 'returns the suffix as an attribute' do
       expect(supporter.suffix).to eq('IV')
     end
   end
 
-  describe "#email" do
-    it "returns the email as an attribute" do
+  describe '#email' do
+    it 'returns the email as an attribute' do
       expect(supporter.email).to eq('johnjacob@example.com')
     end
   end
 
-  describe "#receive_email" do
-    it "returns the receive email as a boolean" do
+  describe '#receive_email' do
+    it 'returns the receive email as a boolean' do
       expect(supporter.receive_email).to eq(true)
     end
   end
 
-  describe "#phone" do
-    it "returns the phone as an attribute" do
+  describe '#phone' do
+    it 'returns the phone as an attribute' do
       expect(supporter.phone).to eq('1234567890')
     end
   end
 
-  describe "#street" do
-    it "returns the street as an attribute" do
+  describe '#street' do
+    it 'returns the street as an attribute' do
       expect(supporter.street).to eq('123 Main St')
     end
   end
 
-  describe "#street_2" do
-    it "returns the street_2 as an attribute" do
+  describe '#street_2' do
+    it 'returns the street_2 as an attribute' do
       expect(supporter.street_2).to eq('Apt 404')
     end
   end
 
-  describe "#city" do
-    it "returns the city as an attribute" do
+  describe '#city' do
+    it 'returns the city as an attribute' do
       expect(supporter.city).to eq('Schnechtady')
     end
   end
 
-  describe "#state" do
-    it "returns the state as an attribute" do
+  describe '#state' do
+    it 'returns the state as an attribute' do
       expect(supporter.state).to eq('NY')
     end
   end
 
-  describe "#zip" do
-    it "returns the zip as an attribute" do
+  describe '#zip' do
+    it 'returns the zip as an attribute' do
       expect(supporter.zip).to eq('12345')
     end
   end
 
-  describe "#country" do
-    it "returns the country as an attribute" do
+  describe '#country' do
+    it 'returns the country as an attribute' do
       expect(supporter.country).to eq('USA')
     end
   end
 
-  describe "#source" do
-    it "returns the source as an attribute" do
+  describe '#source' do
+    it 'returns the source as an attribute' do
       expect(supporter.source).to eq('rspec')
     end
   end
 
-  describe "#status" do
-    it "returns the status as an attribute" do
+  describe '#status' do
+    it 'returns the status as an attribute' do
       expect(supporter.status).to eq('Active')
     end
   end
 
-  describe "#tracking_code" do
-    it "returns the status as an attribute" do
+  describe '#tracking_code' do
+    it 'returns the status as an attribute' do
       expect(supporter.tracking_code).to eq('abc123')
     end
   end
 
   describe 'dates' do
-    it "returns the status as an attribute" do
+    it 'returns the status as an attribute' do
       expect(supporter.date_created.iso8601).to eq('2014-03-14T14:07:29-04:00')
       expect(supporter.last_modified.iso8601).to eq('2014-03-14T13:54:10-04:00')
 
@@ -228,14 +229,14 @@ describe SalsaLabs::Supporter do
     end
   end
 
-  describe ".fetch" do
+  describe '.fetch' do
     let(:supporters_fetcher) { double('SupportersFetcher', fetch: []) }
 
     before(:each) do
       allow(SalsaLabs::SupportersFetcher).to receive(:new).and_return(supporters_fetcher)
     end
 
-    it "calls .fetch on an SalsaLabs::SupportersFetcher object" do
+    it 'calls .fetch on an SalsaLabs::SupportersFetcher object' do
       SalsaLabs::Supporter.fetch
 
       expect(supporters_fetcher).to have_received(:fetch)
@@ -243,14 +244,13 @@ describe SalsaLabs::Supporter do
   end
 
   describe 'save' do
-
     let(:object_saver) { double('SalsaObjectsSaver', save: []) }
 
     before(:each) do
       allow(SalsaLabs::SalsaObjectsSaver).to receive(:new).and_return(object_saver)
     end
 
-    it "calls .fetch on an SalsaLabs::SupportersFetcher object" do
+    it 'calls .fetch on an SalsaLabs::SupportersFetcher object' do
       supporter.save
 
       expect(object_saver).to have_received(:save)
